feat: Support poetry version format >= 1.5.0
We've been blocked by dparse library not being able to parse newest poetry lock versions, which was making our pipelines fail since we run twyn against our dependencies. We will consider completely dropping the usage of dparse, to make Twyn more lightweight.

Since we now require `tomllib`, which is present on Python versions 3.11 onwards, and this tool is now meant to be run standalone (meaning, not as a package), we're also dropping the tests against previous Python versions.

Also added some comments to the workflow files.