feat: treat namespaces as a special case when checking for typos (#378)
This PR makes a difference between normal packages and packages with a
namespace. For the latter, the behaviour has been extended so that we
check only in the namespace for a typo, and if there's a possible one,
we require the package to be identical to the legit one.

For example:

- trusted package -> `@aws/sdk`
- not a typo -> `@awz/zdk` (namespace could be a typo, but package name
does not match trusted one)
- typo -> `@awz/sdk` (namespace could be a typo, package name matches
trusted one)


This will help reduce the number of false positives and assumes that if
someone is impersonating a package, they'll mimic the namespace but keep
the package identical.

refs #97